Include information about location of package if refresh fails
Previously the stack traces when a package failed to load (due to miscellanies reasons) would not include which package that failed, thus making it hard to debug and fixing the error.
This commit fixes this by including the resource description in the stack trace and wrapping the original stack trace.
A test util class for tapping system err is added for the ability to verify the solution, as the exception is just logged and not propagated out of the class.
Coding style is kept as close to the rest of the other files as possible.

Solves issue #228