Add restrictions to catalogs based on user attributes
By providing a restrictions-object to the catalogs.json we can now restrict which catalogs should be presented to the user based on some attribute they have. This may be useful in scenarios where one only wants certain users some catalogs.
From this commit it still remain to implement a check on actual launch, as well at documentation of how the restrictions-object should be configured in the json

Co-authored-by: Olivier Levitt <olivier.levitt@gmail.com>