Set field catalog in CatalogWrapper to a default object
Such that the field exists when the catalogwrapper is returned from the API. This makes the API more consistent, that one always can expect the fields to exist. This will help prevent unnecessary errors for the consumer.
The PR is somewhat kind of extension of the health check (which doesn't return 200 OK before the catalogs are loaded).